+++ src/Server/ServerCore/Systems/ServerChat/ServerChatInternalController.cs
@@ -14,6 +14,7 @@
 // You should have received a copy of the GNU General Public License
 // along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
+using System.Numerics;
 using Sovereign.EngineCore.Components.Indexers;
 using Sovereign.EngineCore.Events;
 using Sovereign.EngineCore.Events.Details;
@@ -90,16 +91,14 @@ public class ServerChatInternalController
     /// <param name="red">Red component of the font color.</param>
     /// <param name="green">Green component of the font color.</param>
     /// <param name="blue">Blue component of the font color.</param>
-    public void SendMessageToPlayer(ulong playerEntityId, string message, byte red, byte green, byte blue)
+    public void SendMessageToPlayer(ulong playerEntityId, string message, Vector3 color)
     {
         var details = new GenericChatEventDetails
         {
             Message = message,
             Target = GenericChatTarget.Player,
             TargetId = playerEntityId,
-            ColorRed = red,
-            ColorGreen = green,
-            ColorBlue = blue
+            Color = color
         };
         var ev = new Event(EventId.Core_Chat_Generic, details);
         eventSender.SendEvent(ev);
@@ -109,18 +108,14 @@ public class ServerChatInternalController
     ///     Sends a generic message to all players.
     /// </summary>
     /// <param name="message">Message.</param>
-    /// <param name="red">Red component of font color.</param>
-    /// <param name="green">Green component of font color.</param>
-    /// <param name="blue">Blue component of font color.</param>
-    public void SendMessageToAll(string message, byte red, byte green, byte blue)
+    /// <param name="color">Text color.</param>
+    public void SendMessageToAll(string message, Vector3 color)
     {
         var details = new GenericChatEventDetails
         {
             Message = message,
             Target = GenericChatTarget.Global,
-            ColorRed = red,
-            ColorGreen = green,
-            ColorBlue = blue
+            Color = color
         };
         var ev = new Event(EventId.Core_Chat_Generic, details);
         eventSender.SendEvent(ev);